requirements.txt
dspy
flask
flask_cors
pdfplumber
transformers
faiss-cpu
langchain
langchain_community
pypdf
openai
matplotlib
numpy
pandas
pytest
dash
plotly
python-dotenv
pymupdf 